chore: self contained examples
