--- examples/demo/mod.rs
@@ -1,3 +1,0 @@
-mod app;
-pub mod ui;
-pub use app::App;
